+++ src/bin/tai64n.rs
@@ -1,0 +1,13 @@
+use std::io::BufReader;
+
+fn main() {
+    if std::env::args_os().len() != 1
+        || rgbdns::tai64::stamp(
+            BufReader::new(std::io::stdin().lock()),
+            std::io::stdout().lock(),
+        )
+        .is_err()
+    {
+        std::process::exit(111);
+    }
+}